+++ ta4j-core/src/test/java/org/ta4j/core/criteria/drawdown/DrawdownTest.java
@@ -21,7 +21,7 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.criteria.helpers;
+package org.ta4j.core.criteria.drawdown;
 
 import org.junit.Test;
 import org.ta4j.core.BaseTradingRecord;
